Rename generate rtd module
